requirements.txt
Django~=4.2
django-environ==0.10.0
djangorestframework==3.14.0
pre-commit==3.3.3
psycopg==3.1.9
dj-database-url==2.0.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
requests~=2.31.0
